Refactor get_config_dir import across modules
The get_config_dir function import has been refactored to be imported from the docuverse.utils module rather than from the elastic submodule in both the search_engine.py and retrieval_servers.py files. Additionally, the get_param function import was also updated in evaluator.py.
